Get rid of split package org.jetbrains.kotlin.incremental in Gradle plugin
    #KT-18621 fixed
